ImageData: Unify creation paths
The goal here is to add an extra parameter to ImageData::
ValidateAndCreate, and in that parameter specify the color space to
use, if the ImageDataSettings doesn't set one.

The function is getting unwieldy, so add a structure of
ImageData::ValidateAndCreateFlags, instead of the bitfield stuff.
Add the color parameter there.

Refactor CanvasRenderingContextBase::getImageData to use ImageData::
ValidateAndCreate to allocate its data array (instead of its current
one-off implementation). The reason for the one-off was because
that function needed a "don't zero initialize" option, which was
easy to put into the ImageData::ValidateAndCreateFlags.

The next step will be to make CanvasRenderingContextBase::getImageData
and createImageData specify a default color space that isn't sRGB.

Bug: 1208480
Change-Id: I61de71f2c95892e785b040190bccbc3d709480be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2893237
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883651}
NOKEYCHECK=True
GitOrigin-RevId: 828691d0e84270a7c62af86ec273b581b2b59a51