Move the whole subsequence subtree when moving a cached subsequence
For the following cached subsequence tree:
  Subsequence A
    Subsequence A1
    Subsequence A2

previously, during a paint, when A could use cache, we moved the paint
chunks and display items from cache for A, but didn't keep A1 or A2.
During the next paint, if something under A2 changed, we had to repaint
the whole A because we didn't have information about A1 or A2.

Now when moving A, keep A1 and A2 so that if one of them changes while
another doesn't, we can use cache for the unchanged one.

This improves performance of record_time_small_invalidation by 25%:
https://ct.skia.org/results/cluster-telemetry/tasks/chromium_perf_runs/wangxianzhu-ChromiumPerf-5812/html/index.html

Bug: 917911
Change-Id: I29dde93c08496a0a9fa507d2db30291552426965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2885012
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883716}
NOKEYCHECK=True
GitOrigin-RevId: aee7fd84efbd6b44ab8b448b8fefefa404bf6a2c