Propagate used writing-mode from html root to text children
The writing-mode and direction is propagated to the html root from the
body element if they differ in computed values. On initial layout tree
attachment that already worked, but we also need to need to propagate
the changed used styles down to text children on style changes when we
don't reattach the layout tree.

Bug: 1209144
Change-Id: Id16d026da49e86a73ebacfb927938896fc4d84b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2897286
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883978}
NOKEYCHECK=True
GitOrigin-RevId: 1e4032eb2bf3cae078f6108d7c9c03215a527c19