Remove CanUpdateLayout.
CanUpdateLayout was added in https://codereview.chromium.org/2762553002/
to ensure there wasn't a behavioral change in the code. Printing is the
only WebViewClient that does not override this value. The return value
of false was only used for printing selections (when a new WebView is
created) and non-selection printing got the value from RenderViewImpl
which was always true.

BUG=1155202

Change-Id: I17c6b9b3426e98d2d474abeb80de381ac0560d8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2879658
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884029}
NOKEYCHECK=True
GitOrigin-RevId: 9abcb39937476d11d2d661e7d24d50cd5697f6d4