Allow CanvasResourceProviderSwapchain to emit OOPR status
Images were not properly serialized in lowlatency mode in the OOPR path.
This was happening because the image provider was unable to determine if
the resource provider was operating in OOPR mode.

This change ensures that the image provider can check if the
CanvasResourceProviderSwapchain is running in in OOPR mode.

Bug: 1093147
Change-Id: I837f8895f0151a455069acfd1fc0490ce1caca30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2901588
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Jonah Chin <jochin@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#884030}
NOKEYCHECK=True
GitOrigin-RevId: cf7c9f49c6ea23952e1be6a05594ff7e7d031dda