Fix WPT css/css-transforms/transform3d-sorting-006.html test.
This changes the test (since it's easier to modify than the reference)
to match both the spec and the interoperable behavior of Chromium,
Gecko, and WebKit.

It still needs to be marked as failing because the antialiasing is
different between test and reference (at least on Linux).

Change-Id: Ifb1c17973175fdd6244d9897ae53d50952638073
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2904125
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884114}
NOKEYCHECK=True
GitOrigin-RevId: 986610e4ffdb4d3ea8ad7aca0b7d16e724f6b826