PlzDedicatedWorker: add a new UseCounter
The existing UseCounters, kWorkerControlledByServiceWorkerOutOfScope and
kWorkerControlledByServiceWorkerWithFetchEventHandlerOutOfScope, count
dedicated workers that will not be controlled by the service worker
that is the same as the ancestor frame is controlled.
These UseCounters, however, overlook dedicated workers that will be
controlled by a different service worker from the ancestor's one after
PlzDedicatedWorker is enabled.

This CL adds a new UseCounter to exclude the cases that a dedicated
worker is controlled by one of service workers registered for the origin.

Change-Id: I34e2c1277fd02e5237fd602ecfa05e4b791589e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891872
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884403}
NOKEYCHECK=True
GitOrigin-RevId: 8b756a4aa3432bd137f54c078e7996d091d8036e