[Private Network Access] Simplify PrivateNetworkRequestPolicy.
The immediate goal of this CL is to unblock CL 2847430. Beyond that, it
helps expand the set of expressible policies from the point of view of
the content layer without adding yet more verbose entries to the policy
enum, and generating their cross product. It is also something I had
been looking forward to doing with preflight requests in mind, where we
will want some kind of kPreflight policy.

This change is a no-op from the browser's perspective. Speculative RFHs
are unaffected too: in the unlikely event that
RFHI::BuildClientSecurityState() is called on a speculative instance
(which can happen due to race conditions, as documented in the method
definition), we ignore RFHI::private_network_request_policy_ anyway.

Bug: chromium:1141824, chromium:1183811, chromium:591068
Change-Id: I4a2e354d2fa761f576782a5d21bdce5332988e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2897985
Auto-Submit: Titouan Rigoudy <titouan@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884431}
NOKEYCHECK=True
GitOrigin-RevId: 74658deabf07caa7450427e3d648ba928fbe5b53