[CaptureHandle][#4] Expose correct CaptureHandle to render process
CL Chain
--------
This is the 4th CL in a chain. The expected length of the *core* of the
chain is 6-7 CLs. Extensions likely to follow. The first CL has a full
explanation of the feature and an outline of the chain:
https://chromium-review.googlesource.com/c/chromium/src/+/2887150/

Contents of this CL
-------------------
In the browser process capture handles are observed.

A captured tab can have its capture handle changed by:
a. Captured application calls setCaptureHandleConfig().
b. Captured application's top-level frame is
   navigated to another document.
c. User clicks the [share-this-tab-instead] button.

When the capture handle of a captured application X changes,
for each capturing application Y, calculate the new observable
capture handle and send it to the capturer's render process.
The next CLs deal with handling this message.

Clarification: CaptureHandleConfigs set permitted-origins.
By "observable capture handle" we mean the capture handle
as-is for allowlisted capturers, and the empty capture handle
for non-allowlisted capturers.

Bug: 1200910
Change-Id: I1a160d01ba21618fbf0dd0e99b7969a316836684
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2897153
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884467}
NOKEYCHECK=True
GitOrigin-RevId: a96657382fc3a71144f27ebd3b37157a018123f5