[CaptureHandle][#5] Update capture-handle returned by getSettings
CL Chain
--------
This is the 5th CL in a chain. The expected length of the *core* of the
chain is 6-7 CLs. Extensions likely to follow. The first CL has a full
explanation of the feature and an outline of the chain:
https://chromium-review.googlesource.com/c/chromium/src/+/2887150/

Contents of this CL
-------------------
Pipe capture-handle updates received on the render-side from
MediaStreamDeviceObserver to WebPlatformMediaStreamSource::device_.
Once there, calls to MediaStreamTrack.getSettings() will pick up
the new value.

Bug: 1200910
Change-Id: I65c5e0ad55fb5ee4718a80b0b48c3d63bc633b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2896319
Commit-Queue: Elad Alon <eladalon@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884470}
NOKEYCHECK=True
GitOrigin-RevId: f7d1779301ed7afff9605b2f149e322673c7ddc3