Avoid unconditional strong references to scroll-linked animations
We currently store strong references to animations that attach to
ScrollTimelines. The are only two ways for an animation to leave the
strong set: (1) if the timeline of the animation is changed, and
(2) via Animation::Disposed (GC finalizer). This is problematic, because
ScrollTimelines are strongly referenced by their associated scroll
source (see Node::RegisterScrollTimeline) while any animation is
attached. Hence, as long we don't switch the timeline or GC the
source Node, scroll-linked animations can *never* be garbage collected
once they are attached to a ScrollTimeline, no matter how hard they are
canceled and forgotten about by JS.

AnimationTimeline already has a system for keeping relevant animations

animations that need updates. This should be sufficient for
ScrollTimeline as well, so we can just remove ScrollTimeline::
scroll_animations_ and use AnimationTimeline::GetAnimations
whenever we need to iterate through the animations attached to the
ScrollTimeline.

alive: a weak set for all attached animations, and a strong set for
Change-Id: I3e3a11567fbc97b4d5fbfae34d62f3e3c5633b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2876787
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884485}
NOKEYCHECK=True
GitOrigin-RevId: 7a527624195ab0fcd9dd8a425b16a6799236fb25