Implement MediaCapabilities API decodingInfo for WebRTC
"webrtc" was recently added as an MediaDecodingType to the
MediaCapabilities API, see
https://w3c.github.io/media-capabilities/

This CL implements support for responding to supported and powerEfficient.
smooth is currently hardcoded to be true given that the configuration
is supported.

The feature is implemented behind the blink runtime enabled flag
MediaCapabilitiesWebRtc

Bug: chromium:1187565
Change-Id: I3bc72a6a8dd06d6bfb84badbcc24f50e77fd1e67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2794472
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884573}
NOKEYCHECK=True
GitOrigin-RevId: 14f4398611991b1e0e8ce1d80c67139c20b0cf71