Minor cleanup of ValidateClapProperty()
Add some comments describing the calculations and the formulas used.

Remove the checks to see if the center of the clean aperture is within
the image's bounds, because they are implied by the checks to see if the
four corners of the clean aperture are within the image's bounds.

Bug: 1167422
Change-Id: Ib32e20a15bcd994e45a8bad79913a8a578270016
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2906662
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/master@{#884693}
NOKEYCHECK=True
GitOrigin-RevId: cff3d86c51e7d94247c78162f9736d3c0be07c76