WebTransport: SendStream inherits WritableStream
Update SendStream to be a subclass of WritableStream, matching the
latest version of the WebTransport standard.

The old SendStream APIs are retained for backwards-compatibility. The
"writable" attribute returns the same object so that old code works.

The code now detects errors during creation of the WritableStream and
attempts to handle them.

WritableStream::CreateWithCountQueueingStrategy now uses a trivial
queuing strategy, the way ReadableStream's version has always done. This
was done to reduce the possibility of errors during construction, but
may give a minor performance improvement.

BUG=1011392

Change-Id: I790dfe7978c167fdab8090b00232475b20de94f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905254
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#884867}
NOKEYCHECK=True
GitOrigin-RevId: 108f5defe468f2eb5f6d9783cb9ec4173f37b287