Push BackgroundImageGeometry down into BoxPainterBase helpers
Rather than 'exploding' the tiling parameters we can just pass the whole
object a bit further. This allows additional computations to be
performed in LayoutUnit space, and may allow refining some of the
heuristics in the future.
Also don't pass the destination offset to ComputePhaseForBackground()
since it's only added to the computed phase - something that can be
done in the caller. Similarly only return an offset since the size will
always be the tile size. Since all the inputs to the function is now
LayoutUnits, the computations is done using LayoutUnit-appropriate
functions.

Bug: 1208044
Change-Id: I33303507fa6be526d8d8c3d96030b2622f8eb065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905535
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#884915}
NOKEYCHECK=True
GitOrigin-RevId: fd6bd8fe0d2805620d886ee31f8129264f511b59