Rename the SpeculationRules trial to SpeculationRulesPrefetch.
Most use of the trial will not be proxied, so this name is now
misleading.

The REF and base::Feature are not renamed in order to keep this
change as simple and safe to merge as possible.

Bug: 1211416
Change-Id: Ia383802259a9edda44c388940aee1a69e8e50c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910852
Auto-Submit: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Jason Chase <chasej@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885186}
NOKEYCHECK=True
GitOrigin-RevId: 7176d93ab99369b869f6843f6b9a4cd650cc5db0