Add plugin loading delay to test, to reduce flakiness
Since we can't listen for an event [1] that indicates the plugin has
loaded, we'll just give a bit more time for it to happen. This test is
flaky because sometimes the plugin loads and generates the console
message:

  CONSOLE MESSAGE: Blink Test Plugin: initializing

and sometimes it doesn't. This added delay should reduce flakiness.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ppapi/tests/blink_test_plugin.cc;l=42;drc=4e63a512000230e68d72103056abdaf38b1f019b

Fixed: 1208303
Change-Id: Ia6d476ed2df574ae9a2427ef76f787fdc2f712df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910940
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885255}
NOKEYCHECK=True
GitOrigin-RevId: 7c5975e691c455257ca1d86db0382c43c023bcff