Change long-sibling-list.html to a measurement-time test
This test is quite noisy, likely because there was an O(N^2)
nature to the test. It was previously appending continuously
to the child list, so measureRunsPerSecond() would get slower
with each iteration. The new test measures the time for a
fixed length (30000) list of children. It also asserts that
the garbage collection APIs are present. Without an actual
GC [1], the test is also quite noisy.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/perf_tests/resources/runner.js;l=106;drc=c5cb097858b17e5fb7a37d64dd5cf908c4a13da2

This will mark the regression fixed, because the new metric
will be in different units:

Fixed: 1209631
Change-Id: Ie97e0af0f5455d13aec3a1ba04e4b711904f4913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2909998
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885307}
NOKEYCHECK=True
GitOrigin-RevId: 9df38067498a07f9b642313d90f2753615f1c08b