[LayoutNG] Multicol writeback to legacy layout
Only clamp the fragmentainer block size from 0 to 1 in the case
where the fragmentainer content overflows the zero-height
fragmentainer when writing back to legacy. This more closely matches
what legacy is expecting. For example, in the case where a spanner
breaks its parent content, resulting in a zero-height column before
the spanner, legacy expects this column's height to be 0 instead of
clamped to 1. This update prevents virtual/layout_ng_block_frag/
external/wpt/css/css-multicol/filter-with-abspos.html from failing in
CL:2876245.

Because the clamping behavior for fragmentainers may be different
between legacy and NG, we need to store an extra adjustment in the
associated NGBlockBreakToken to ensure the correct consumed block size
for the fragmentainer is used when writing back to legacy.

Bug: 1210287
Change-Id: I0f067b10bd1d0e31b21b676e9c28fa8b1e9fba29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2904547
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#885308}
NOKEYCHECK=True
GitOrigin-RevId: af944599d85336a44114c338f209ba3387542e1c