[Keyboard Tooltip] Mitigate bug with keyboard-triggered tooltips in OOPIF
While working on CL:2844544, I noticed that HasStickyActivation returns
false when the user moves the focus inside of an OOPIF with the tab key,
before any other user action event. This was preventing
keyboard-triggered tooltips to be triggered when at the start of an
OOPIF.

I filed https://crbug.com/1208874 and folks who knows more than about
sticky activation will investigate. In the meantime, here's a mitigation
that should be safe to use.

A test is added in the aforementioned CL.

Bug: 829352
Change-Id: Idff321010dbda4a0b17cc4ea4ef21dc45448d801
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2908121
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#885491}
NOKEYCHECK=True
GitOrigin-RevId: c062c62f11e87cc76d4c68a10b755c31cd58c564