[LayoutNG] OOF contained by inner multicol
An OOF contained by an inner multicol was not getting laid out because
we skipped propagating up out-of-flow descendants when we reached
an inner multicol (because if the containing block was somewhere inside
the inner multicol, we would lay out those OOFs once we reached the
outer multicol). However, we should not skip propagating the OOF
descendant if the inner multicol is the containing block.

Bug: 1211522
Change-Id: I00876c44eeaa415edd6a90d1d29712077a2ce24c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2911733
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885562}
NOKEYCHECK=True
GitOrigin-RevId: d7d078dc1d52fd5eb699b2d3d1a170633dd8e8de