Adding parameter to specify when we want willdraw in canvas
Changing the Canvas functionality to include a WillDraw was properly
fixing the video issue, but it was creating some side effects in other
places where Snapshot would be called afterwards.

To avoid this issues and improve future changes, this CL adds a new
parameter to just use the willDraw whenever it's needed inside Canvas.
The caller will have to specify if they need it or not.

Two follow-up issues will be to investigate whether this memory leak is
something we can prevent in any other way, and to add a new perf test
that will draw several patterns to a Canvas, as the bug 1211181 repro
case suggests.

Bug: 1211181
Change-Id: I3b96e58555940123be6e5c4495c4442fcb4e7a50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910706
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885599}
NOKEYCHECK=True
GitOrigin-RevId: 2186225df07e1dd21e8424cf925f14a7cb349a94