IsSlotAssignmentRecalcForbidden() check is not necessary
Now that there is an early return for <slot> (and thus
FlatTreeTraversal is not used on it), this check is not necessary
and actually impedes cleanup when layout objects are being destroyed.

A dependent change, CL:2903990, requires this because of some DCHECKs
it adds that ensure image children are not pointing to a detached
parent. Without this change, the function cannot continue to use the
DOM for cleanup purposes.

Bug: None
Change-Id: I218be4e0de189b6c1c63a7dd47c216bd466bef9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2909875
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885658}
NOKEYCHECK=True
GitOrigin-RevId: c36d41efd12cfde7473a4b2e52014248ba670bb8