[a11y] System font sizes should obey user's font size setting
Some form controls get assigned a system font in UA origin:

  input, textarea, select, button {
      font: -webkit-small-control;
  }

The font size of that is supposed to be 2pt smaller than the default
font size. However, that "default" was hardcoded to 16px, without taking
into account that the user can change the actual default.

For example, someone with suboptimal vision may use a bigger default in
order to be able to read the text with ease. But form controls would
ignore that preference. This was bad for accessibility.

So this patch uses the actual default size when calculating the size
of system fonts.

Doing so requires a Document from which to retrieve the settings, but
CSSParserContext::GetDocument() is null for UA stylesheets. Therefore,
instead of resolving at parse time, the font longhands are set to a
CSSPendingSystemFontValue, and their resolution is deferred until
computed-value time, when we can use StyleResolverState::GetDocument().

Bug: 1054031

TEST=StyleEngineTest.SystemFontsObeyDefaultFontSize
TEST=external/wpt/css/css-fonts/system-fonts-serialization.tentative.html
TEST=fast/css/font-shorthand.html

Change-Id: I5e9b41a2baa45cea7f1a54190bf593e025153adb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891582
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885715}
NOKEYCHECK=True
GitOrigin-RevId: 246e51638dab47ece909855ab35be2661bd98c1f