[Keyboard Tooltip] Allow keyboard triggered tooltip from the renderer
In this CL, we finally connect the browser-side implementation for
keyboard-triggered tooltips to the renderer-side trigger. Now, using the
newly added WidgetHost::UpdateTooltipFromKeyboard, we will be able to
trigger a tooltip by setting the focus on a web content control from the
renderer process using the keyboard.

CL:2866730 added the trigger on the renderer side.

Bug: 829352
Change-Id: I46eb7120d8fc24a017442ade1e35f57cb001d472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2844544
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#885725}
NOKEYCHECK=True
GitOrigin-RevId: 036dcd86f523c6cb7f6b50eac11b8bee579c76c2