Introduce [Web]LoaderFreezeMode
Define WebLoaderFreezeMode and replace WebURLLoader::DeferType with it,
and define LoaderFreezeMode, a typedef for the enum in blink/renderer.

This is part of the defer/freeze naming cleanup, and
[Web]LoaderFreezeMode and its members are better names than DeferType
and its members.

This is also good for build time because this reduces the include
directives for public/platform/web_url_loader.h.

Bug: 1211696
Change-Id: Ie4272ded8c25b3c59f382c53bb8f7f36fe315bd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2912337
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885937}
NOKEYCHECK=True
GitOrigin-RevId: 731ba54eecd14ac1cbb36eafb24053bb500cc3aa