URLPattern: Add a lenient liburlpattern tokenizer policy.
Create a lenient Tokenizer policy that causes liburlpattern::Tokenize()
to return a list of Tokens with a kInvalidChar type instead of returning
an error.

This will be used in future CLs to parse a URLPattern constructor string
that combines multiple liburlpattern patterns together.  The lenient
tokenizer policy will be used to find potential separators between these
patterns while avoiding the details of understanding named groups,
regex groups, etc.

For example, in the URLPattern constructor it is planned to use this to
help parse constructor strings like:

  new URLPattern("https://example.com/:name?/foo?bar");

Using the lenient tokenizer mode allows us to easily differentiate
special characters like the colon in `://` and the question mark in
`?bar` as being not part of a real pattern.  Meanwhile it is obvious
that `:name?` is part of a real pattern because it tokenizes as a
TokenType::kName followed by a TokenType::kOtherModifier.

The lenient mode should make it easier to split the string into URL
component patterns.

Bug: 1141510
Change-Id: I0a51a376313234df75302db6711a52768fba9edd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2902238
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885958}
NOKEYCHECK=True
GitOrigin-RevId: 6ef4d3fb3dcb4e85a8ca43092ee8e774b48211d3