[GridNG] Change <meter> element to stop relying on a grid bug.
Previously legacy grid would only consider items on the grid row as
contributing to the baseline.

The <meter> element was exploiting this fact as it placed an inflow
element on the 2nd row - no contributing to the baseline.

GridNG doesn't have this bug. It'll look for the first non-empty row,
and synthesize the baseline from that.

An example of this is:

baseline <div style="
    border: solid;
    display: inline-grid;
    width: 20px;
    grid-template-rows: 10px 20px 10px;
">
  <div style="grid-row-start: 2; background: cyan;">

This change the internals of the grid to use an absolute child instead.
This isn't inflow - and as such doesn't contribute to the baseline.

Bug: 1045599
Change-Id: I4a6f8a975cf726414026c44d6e0289722a807e76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2909363
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885961}
NOKEYCHECK=True
GitOrigin-RevId: a7bfa3ea3ce60df9ecece20472ef9eac37b35f2e