Make WebRTC-in-Chrome's libyuv scaling filter match WebRTC.
kFilterBox is the best quality but slower. kFilterBilinear is faster
but lesser quality.

Chrome has been using kFilterBilinear for some time now and WebRTC has
been using kFilterBox. But as it turns out, WebRTC-in-Chrome has been
using WebRTC's scaling function (= box) for lower simulcast layers so
saying that "Chrome uses bilinear" was not true in a lot of cases. This
accidentally changed after some refactoring in M91 where CropAndScale
was called prior to mapping the frame instead of afterwards. So Chrome
currently uses a lesser quality than standalone WebRTC and this is to a
large extent a quality regression. (Happens in VP8 but not VP9.)

To fix the quality regression and get the different code paths to agree
on what filter to use, this CL changes the Chromium scalers used by
WebRTC to also use kFilterBox.

In common cases this makes little CPU difference (e.g. 2 and 4 are fast
paths) but for other scaling factors like 2.5 kFilterBox may cost us
1.0% more CPU according to profiling. This is not ideal, but we have
old bug reports documenting that kFilterBilinear causes unacceptable
amounts of aliasing (b/152837511) so this is worth it for the sake of
quality.

On the plus side, recent changes allow us to skip some unnecessary
downscales, which means we can skip 0.1-1.0% CPU in "Brady Bunch"
scenarios.

If we want to challenge the quality vs performance tradeoffs of the two
filters we can add an experiment for this in the future, but for now we
need the two code paths to agree on filter and to fix the quality
regression.

Bug: chromium:1212630
Change-Id: I72ae69e749e168872c595669e0a25adf2a08b70b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910627
Reviewed-by: Evan Shrubsole <eshr@google.com>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#885966}
NOKEYCHECK=True
GitOrigin-RevId: 2f88586c48ed6c1daf3bf23eee28697fd1f6a86c