[baselines] Fix synthesising vertical baselines for flex.
Previously we always synthesized baselines (outsize an inline formatting
context) using the block-end edge.

This changes the behaviour to use the middle if in a vertical writing
mode. This isn't "complete" as it is still missing the text-orientation
logic, but significantly better.

Bug: 1076080
Change-Id: I14f5aea5adb4f3afcd9194b6f4f9b54addeb54ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905738
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886062}
NOKEYCHECK=True
GitOrigin-RevId: b3e8f319962d1c03c36cac88923665f09f3b3c0e