cancelScheduledValues throws RangeError for negative values
The spec clearly states that cancelScheduledValues must throw a
RangeError for negative values.  Make it so, and add a test for this.
We also add tests for NaN and Infinity.

Note, however, that the spec says RangeError is also thrown for
non-finite values.  This is incorrect since cancelTime is a double,
not unrestricted double.  (See
https://github.com/WebAudio/web-audio-api-v2/issues/127)

Thus, the test throws TypeError for NaN and Infinity.

Bug: 1212307
Change-Id: I750ab0f2b9f17f94d04cb2e9422b4f626252dfbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2909615
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886277}
NOKEYCHECK=True
GitOrigin-RevId: 704beb7c52be6be6559ac8fe7e28f72130b653bb