Fix preemptive mapping of DXGI buffers
Previously, a shared memory region was created and filled with pixel
data during capture if requested by a client. But that doesn't work with
current video capture service design: first, a GMB handle is passed once
and then the client is notified when a frame is available in that handle.

To make preemptive mapping work with that design, a shared memory region
is now always created together with the DXGI GMB handle in the capture
service. Then, when a frame is captured and preemptive mapping is needed,
the region is filled with data and a special flag is passed to the
consumer in VideoFrameInfo. Based on that flag a previously received
shared memory region is passed to the frame buffer.

Bug: 1120900
Change-Id: I6f55be94bab76e9f2d8f5ec0e3b7f7e8badb8fa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905782
Commit-Queue: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886340}
NOKEYCHECK=True
GitOrigin-RevId: 3d1691f5c500ba497f36bc3af7bb0e181bb2ac50