scheduler.postTask on workers platform/scheduler changes.
This CL contains the plumbing changes needed in the scheduler for
putting scheduler.postTask on workers as well as code that defines how
prioritization in workers will work (for now it is the same as the main
thread scheduler). It will be followed up with changes in
modules/scheduler which will expose the API on workers. See
https://chromium-review.googlesource.com/c/chromium/src/+/2864467 for
the full picture.

Change-Id: I2edab71c695ee29d0c878ce46c086cb0e37019a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2911513
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886450}
NOKEYCHECK=True
GitOrigin-RevId: 400019bf9107e8cf179a73b191f21746190583db