Preload background-image resources in tests
Preload and wait for background-image resources to finish loading before
triggering reftest screenshots. Background images are triggered to load
the first time we do a style resolution which means it is not guaranteed
to be loaded and painted on first paint.

These tests are flaky, and increasingly so with [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2871469

Bug: 332189
Change-Id: Ibefbd75f22a9783e27838ec549072a58ede31325
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2916998
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886469}
NOKEYCHECK=True
GitOrigin-RevId: 8d0ce4739642410a47479df97221eaf591df99a0