Document and test HashSet::find() and erase().
I could not figure out whether HashSet::erase() was undefined if the key
was not found. It turns out it is a no-op, but there were no tests for
this case*, so I added documentation and a test.

Note: One may consider this documentation redundant since this is close
to being an implementation of std::unordered_set, but the interface is
different -- HashSet::erase() returns void whereas
std::unordered_set::erase() returns the number of erased elements, so it
was very unclear whether HashSet::erase() actually has the same
semantics.

*I tested making erase() crash if key is missing, and wtf_unittests
still passes; the new test I added fails in that case.

Change-Id: I431537f1150f7a99924962416b84c37ebbc4cfe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2916775
Commit-Queue: Matt Giuca <mgiuca@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886606}
NOKEYCHECK=True
GitOrigin-RevId: 1c1df7018ba7ac58b5873b1f6715f9061ad9eb8f