EditContext: disable DOM mutation for keyboard shortcuts
This CL disables default DOM mutation triggered by keyboard shortcuts
when EditContext is active.

From EditContext's point of view, there are 3 kinds of commands:
1) BeforeInput event only, ex ctrl+B or <enter>.
2) BeforeInput event + EditContext behavior, ex. backspace
3) BeforeInput event + default DOM behavior, ex. caret navigation.
(this will be no-op for EditContext)

The above logic is implemented in EditorCommand::Execute.

Tests for ctrl+B/I/U are also added to verify the DOM is not
modified and the beforeinput event is fired as expected.

Tests for other commands will be in separate CLs.

Note if the command is from Javascript, i.e. through execCommand,
it's by design that EditContext won't block the DOM mutation since
it's not from user input.

Bug: 999184
Change-Id: I5742a4670b3948bbf749d3c3f0d69bc4db4d2741
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910037
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Commit-Queue: Alex Keng <shihken@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#886643}
NOKEYCHECK=True
GitOrigin-RevId: 68da87262b24d721190140e1b653ba573a800013