[code cache] Make CodeCacheHost interface per-navigation
This cl is in preparation to migrate code caches to use network
isolation key (NIK). GeneratedCodeCaches use response_time provided by
HttpCaches to validate the cached code. Hence it is good to keep
the keys used by HttpCaches and GeneratedCodeCaches in sync. To use
NIK for CodeCacheHost we need to move the processing of these requests
from per process to per-frame. This is first in the series of CLs to
achieve this. More details are in this doc:
https://docs.google.com/document/d/1WRAffSAqdBEfSs6LSZ6Ek8nS35qgjCKmP08-shHxLDE/edit#heading=h.qdjsqrjxndcg

This cl moves the main thread fetch requests for generated code cache
to be processed by RenderFrameHostImpl. More specifically this CL:
1. Update RenderFrameHostImpl to maintain a UniqueReceiverSet for
CodeCacheHost requests. Unlike RenderProcess we must allow multiple
CodeCacheHost interfaces when two navigations are in
progress (for ex: the time between RenderFrameHost sending a commit
message and RenderFrame actually committing).
2. Adds the CodeCacheHost remote interface to DocumentLoader since the
 lifetime of the interface is per-navigation.
3. Implements a way to disable code caches for tests.

Change-Id: I119260305aac0c22632fb60532be0f3264a7d596
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2772986
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Mythri Alle <mythria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886647}
NOKEYCHECK=True
GitOrigin-RevId: fefc0c466da67b3bb39137206963bd0d5fd73729