NGPhysicalBoxFragment flag for fragmentation context root.
We'd incorrectly return true from
NGPhysicalFragment::IsFragmentationContextRoot() for a fieldset with
columns specified, even if the multicol container in such cases is
established by the anonymous wrapper child, not on the fieldset itself
(since the legend shouldn't be part of the fragmentation).

The fragment builder already had a flag for this, which is set by the
column layout algorithm. Store it on the fragment.

Discovered when attempting to land
https://chromium-review.googlesource.com/c/chromium/src/+/2879018
without
https://chromium-review.googlesource.com/c/chromium/src/+/2875086

Bug: 829028
Change-Id: Id017bcd9d2de21c2453b7b9272486110935f0a90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2886424
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#886658}
NOKEYCHECK=True
GitOrigin-RevId: 4c8da6653f7f1991e7cf7c743e9d5845509c2636