Split some COOP WPT tests
These tests seem flaky on Android so lets try to split them as there are
quite a lot in the same test suite that all get executed at the same
time in the same browser session.
This makes them pass on WebLayer on the bots but Clank is still flaky.

Bug: 1175190
Change-Id: I626c66f8ebdabb6d56ebda345d8692d799e94fd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910915
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886800}
NOKEYCHECK=True
GitOrigin-RevId: e4f98b264fa53f1dd93f25b4d120dfedbdc49908