Change the supported H264 profile level in RTCVideoDecoderFactory to 3.1
RTCVideoDecoderFactory::GetSupportedFormats() returns a list of
supported codec profiles. For H264 this also contains information
about the supported level. The previous level that was returned
by RTCVideoDecoderFactory::GetSupportedFormats() was hard coded
to 2.1 (which corresponds to 640x480 at 30 fps).

This CL increases the maximum supported profile level to 3.1
which is the same that is returned by the internal H264 software
decoder in WebRTC.

Bug: chromium:1213437
Change-Id: I99ed9cc85cf09074a7158d4191f8166ad5a9e031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2919811
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886840}
NOKEYCHECK=True
GitOrigin-RevId: fbbba378b78461f65443d5307b6c0666d3ad786c