[Sharing] allow for multiple share calls on clank
This is a partial revert of
https://chromium-review.googlesource.com/c/chromium/src/+/2425933.
The code used to allow for multiple share calls, but that CL changed
it to a single object. This is changing it back, while still enforcing
a single share call except on Android where it's a problem.

Bug: 1096052
Change-Id: I2b677004805215e593fd52458db4a2e57153a739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2915003
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886962}
NOKEYCHECK=True
GitOrigin-RevId: 71003be7ce59254518062bb7fa11ba4dc5106f0b