Extend GetSourceImageForCanvas() with Alpha Disposition param
ExternalImage is defined in WebGPU which includes HTMLCanvasElement,
OffscreenCanvas and ImageBitmap. CanvasImageSource is the base type.
It provides GetSourceImageForCanvas() to return the back resource but
default it need to be applied premultiply alpha op(e.g. ImageBitmap).

This CL extend GetSourceImageForCanvas() with an AlphaDisposition param
to config whether we need some alpha premultiplied/unpremultipled ops
when we get the source image.

For all the types both ExternalImage and CanvasSourceImage have, the
valid value of this alpha_disposition param could be kPremultiplyAlpha
and kUnpremultiplyAlpha. For the remain override implementation, the
valid input must be kPremultiplyAlpha. This CL use DCHECK to guard this.

BUG=1197369

Change-Id: Ia2171ddb66720a429db6befa27c00a8da5e50055
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2908712
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Shaobo Yan <shaobo.yan@intel.com>
Cr-Commit-Position: refs/heads/master@{#886969}
NOKEYCHECK=True
GitOrigin-RevId: 0f715244bfe4376332ed1b7fbbddfa11d1725b24