[devtools] Report issues for invalid SameParty attribute usage
The SameParty attribute can only be used when SameSite=Strict is not
present and the cookie is also marked as Secure. This CL adds an
issue that is filed when SameParty is held wrong.

No CanonicalCookie can be created from a cookie line with wrong
SameParty attribute usage. This means DevTools instrumentation is
also not able to create an AffectedCookie since it relies on
a successful parse of the cookie. To solve this, we add the
raw cookie line to the issue and mark `AffectedCookie` as
optional.

R=chlily@chromium.org, sigurds@chromium.org

Bug: chromium:1179186
Change-Id: I4cfddd87517c03ad2df03f14f08194762245db17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2902705
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#887056}
NOKEYCHECK=True
GitOrigin-RevId: 657178ec55a6242c4149e264ae57654dd6896fd2