[GridNG] Compute SetGeometry after resolving avail. block-size if needed
We (eagerly) compute the SetGeometry when resolving the used track-sizes

This is fine - but in some cases (if we have an indefinite available
block-size initially) we need to recompute once we have a "resolved"
available block-size.

This does this recalculation if we have a non-initial "align-content".

This is "slightly" too broad, e.g. will redo this for
"align-content: start" but the most common case.

Bug: 1045599, 1204689
Change-Id: I150793b2b33c76bef1212ff34509347b3d584338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2919494
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#887227}
NOKEYCHECK=True
GitOrigin-RevId: 91fd48796ae0edb823e593e503d699ced219483c