Include custom properties on computed CSSStyleDeclaration (behind flag)
Since we need a "stable" list of custom properties for length()
and item(), and we also need not-abysmal performance when calling
those functions, this CL caches a vector with the variables names
on ComputedStyle itself.

In CSSComputedStyleDeclaration::item(), we check if the incoming
index is in the range of the standard properties, and if so return
the appropriate one. Otherwise, we're in the variable range (which
is defined per spec to appear after the standard properties), and
we'll fetch the list of variable names.

Unfortunately the inclusion of custom properties requires a clean
style, which means length()/item() now updates style/layout as needed.
This might cause performance regressions, but I don't see a way around
this.

Note: ComputedStyle::StyleInheritedVariables/NonInheritedVariables
were changed to return const pointers, as to not "leak mutability",
which would have made it hard to invalidate the cache. (Nothing
requires those return values to be non-const after StyleCascade
anyway).

Note: ComputedStyle::GetVariableNames() has pretty good test coverage
in ComputedStyleTest already. Added a couple of new ones that target
cache invalidation.

Note: This is a reland-behind-a-flag of a previous CL that was
reverted to due performance regressions.

Bug: 949807
Change-Id: Ieacd5bca1c292ad7f8e126ed113746f8637a9ade
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910120
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#887294}
NOKEYCHECK=True
GitOrigin-RevId: 9ab4966f2a3d0fa31c63a1c24bf365971e91abf4